refactor: move actions to store :)
