+++ src/widgets/BhandaraCheckin/store/actions/index.ts
@@ -1,3 +1,3 @@
 export * from "./resetAppState";
-export { updateDetailsCheckin } from "../../SectionUpdateDetails/updateDetailsCheckin";
+export { updateDetailsCheckin } from "./updateDetailsCheckin";
 export { startCheckin } from "../../SectionMain/startCheckin";
