+++ src/widgets/BhandaraCheckin/store/actions/updateDetailsCheckin/checkinAbhyasi.ts
@@ -1,11 +1,11 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
-import { RootState, ThunkApiConfig } from "../../store";
-import { snackbarSlice } from "../../../../components/Snackbar/snackbarSlice";
-import { postAttendance } from "../../store/api-async-thunks";
-import { UserDetails } from "../../types";
-import { RefinedCityStateCountryLocation } from "../../../../components/LocationTextField/locations";
-import { updateDetailsSectionSlice } from "../updateDetailsSectionSlice";
-import { bhandaraCheckinSlice } from "../../BhandaraCheckin/bhandaraCheckinSlice";
+import { RootState, ThunkApiConfig } from "../..";
+import { snackbarSlice } from "../../../../../components/Snackbar/snackbarSlice";
+import { postAttendance } from "../../api-async-thunks";
+import { UserDetails } from "../../../types";
+import { RefinedCityStateCountryLocation } from "../../../../../components/LocationTextField/locations";
+import { updateDetailsSectionSlice } from "../../../SectionUpdateDetails/updateDetailsSectionSlice";
+import { bhandaraCheckinSlice } from "../../../BhandaraCheckin/bhandaraCheckinSlice";
 
 const getEmailValue = (userDetails: UserDetails): { email?: string } => {
   if (userDetails.email.value?.match(/\*/)) {
