+++ src/widgets/BhandaraCheckin/store/actions/updateDetailsCheckin/checkinMobileOrEmailUser.ts
@@ -1,9 +1,9 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
-import { snackbarSlice } from "../../../../components/Snackbar/snackbarSlice";
-import { UserDetails } from "../../types";
-import { postAttendance } from "../../store/api-async-thunks";
-import { RootState, ThunkApiConfig } from "../../store/index";
-import { bhandaraCheckinSlice } from "../../BhandaraCheckin/bhandaraCheckinSlice";
+import { snackbarSlice } from "../../../../../components/Snackbar/snackbarSlice";
+import { UserDetails } from "../../../types";
+import { postAttendance } from "../../api-async-thunks";
+import { RootState, ThunkApiConfig } from "../../index";
+import { bhandaraCheckinSlice } from "../../../BhandaraCheckin/bhandaraCheckinSlice";
 
 const getEmailValue = (userDetails: UserDetails): { email?: string } => {
   if (!userDetails.email.value) return {};
